services/vision-service/requirements.txt
# TensorFlow Core
tensorflow>=2.15.0
tensorflow-hub>=0.15.0
tf-models-official>=2.15.0

# Model Optimization (quantization, pruning for TFLite)
tensorflow-model-optimization>=0.7.5

# Hyperparameter tuning
keras-tuner>=1.4.0

# Image processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
structlog>=24.1.0
httpx>=0.26.0

# AWS S3 for model storage
boto3>=1.34.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
